Extract player color tonal helpers into colorTone.
Keep diplomacy membership and resolve policy in playerColor; move HSL/RGB tone math to a dedicated module that reuses cartography hex parsing.
